pyproject.toml
[project]
name = "manylatents"
version = "0.1.0"
description = "Unified dimensionality reduction and neural network analysis framework"
license = "MIT"
authors = [
    { name = "César Miguel Valdez Córdova", email = "cesar.valdez@mila.quebec" },
    { name = "Matthew Scicluna", email = "matthew.scicluna@mila.quebec" },
    { name = "Shuang Ni", email = "shuang.ni@mila.quebec" },
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
keywords = ["dimensionality-reduction", "embeddings", "manifold-learning", "pytorch-lightning", "hydra"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core scientific stack
    "numpy>=1.25",
    "scipy>=1.8,<1.15", # upper bound for archetypes/phate compatibility
    "pandas>=2.0",
    "scikit-learn>=1.2",
    "matplotlib>=3.8",
    # Deep learning
    "torch>=2.3,<3.0",
    "lightning>=2.5",
    "torchmetrics>=1.6",
    # Hydra configuration
    "hydra-core>=1.3",
    "hydra-zen>=0.13",
    # Dimensionality reduction algorithms
    "phate>=2.0",
    "umap-learn>=0.5",
    "opentsne>=1.0",
    "multiscale-phate",
    "archetypes>=0.6",
    # Graph tools (used by phate/graphtools)
    "graphtools>=2.0",
    # Utilities
    "tqdm>=4.67",
    "rich>=13.9",
    "Deprecated>=1.2",
]

[project.optional-dependencies]
# Experiment tracking (wandb)
tracking = ["wandb>=0.19"]
# HuggingFace trainer support
hf = ["transformers>=4.40"]
# Neural ODE / SDE dynamics
dynamics = ["torchdiffeq>=0.2", "torchsde>=0.2"]
# Optimal transport (DiffusionMerging)
transport = ["POT>=0.9"]
# Persistent homology / topological metrics
topology = ["ripser>=0.6"]
# SLURM cluster submission
cluster = ["hydra-submitit-launcher>=1.2"]
# GPU-accelerated DR via TorchDR
torchdr = [
    "torchdr>=0.3,<0.4",
    "faiss-cpu",
]
# Leiden community detection for clustering analysis
clustering = ["leidenalg>=0.10", "python-igraph>=1.0"]
# Documentation building
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.6",
    "mkdocs-macros-plugin>=1.0",
]
# All optional extras (excluding docs/dev)
all = [
    "manylatents[tracking,hf,dynamics,transport,topology,cluster,torchdr,clustering]",
]

[project.scripts]
manylatents = "manylatents.main:main"

[project.urls]
Homepage = "https://github.com/latent-reasoning-works/manylatents"
Documentation = "https://latent-reasoning-works.github.io/manylatents/"
Repository = "https://github.com/latent-reasoning-works/manylatents"
Issues = "https://github.com/latent-reasoning-works/manylatents/issues"
Changelog = "https://github.com/latent-reasoning-works/manylatents/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["manylatents"]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=8.3",
]

[tool.uv.sources]
manylatents-omics = { git = "ssh://git@github.com/latent-reasoning-works/manylatents-omics.git" }
shop = { path = "../shop", editable = true }

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "manylatents"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[[tool.uv.dependency-metadata]]
name = "archetypes"
version = "0.11.1"
requires-dist = ["scipy>=1.8,<1.15"]
